+++ src/features/formatter.ts
@@ -1,0 +1,9 @@
+import * as vscode from "vscode";
+
+import { FormattingProvider } from "./providers/format";
+
+export class FormattingEditProvider {
+  activate(): vscode.DocumentFormattingEditProvider {
+    return new FormattingProvider();
+  }
+}